requirements.txt
gorilla-core==0.2.7.8
opencv-python==4.6.0.66
numpy==1.23.1
scipy==1.11.2
open3d==0.18.0
timm==0.6.11